Initial commit
This file cachematrix.R now contains the modified code. The functions performed by this code are included in the R script as well. Here is a brief description of this initial commit:

1. This R file consists of 2 functions : 

a. makeCacheMatrix which creates a special matrix that computes the inverse of a user defined matrix. It has a list of functions that set and get the matrix and set and get the inverse of the matrix.
b. cacheSolve which first checks to see if the inverse has already been computed-if yes, then it returns this value from cache. If not, it computes the matrix inverse using solve(x) function. 

This way, it saves the computational cost for matrix inversion.